[INLONG-11028][SDK] Transform SQL support UrlEncode & UrlDecode functions
