+++ inlong-sdk/transform-sdk/src/main/java/org/apache/inlong/sdk/transform/process/function/string/UrlEncodeFunction.java
@@ -15,30 +15,31 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sdk.transform.process.function;
+package org.apache.inlong.sdk.transform.process.function.string;
 
 import org.apache.inlong.sdk.transform.decode.SourceData;
 import org.apache.inlong.sdk.transform.process.Context;
+import org.apache.inlong.sdk.transform.process.function.TransformFunction;
 import org.apache.inlong.sdk.transform.process.operator.OperatorTools;
 import org.apache.inlong.sdk.transform.process.parser.ValueParser;
 
 import net.sf.jsqlparser.expression.Function;
 
-import java.net.URLDecoder;
+import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 
 /**
- * UrlDecodeFunction
- * description: Decodes a given string in ‘application/x-www-form-urlencoded’ format using the UTF-8 encoding scheme.
- * If the input is NULL, or there is an issue with the decoding process(such as encountering an illegal escape pattern),
- * or the encoding scheme is not supported, the function returns NULL.
+ * UrlEncodeFunction
+ * description: Translates a string into ‘application/x-www-form-urlencoded’ format using the UTF-8 encoding scheme.
+ * If the input is NULL, or there is an issue with the encoding process,
+ * or the encoding scheme is not supported, will return NULL.
  */
-@TransformFunction(names = {"url_decode"})
-public class UrlDecodeFunction implements ValueParser {
+@TransformFunction(names = {"url_encode"})
+public class UrlEncodeFunction implements ValueParser {
 
     private final ValueParser stringParser;
 
-    public UrlDecodeFunction(Function expr) {
+    public UrlEncodeFunction(Function expr) {
         stringParser = OperatorTools.buildParser(expr.getParameters().getExpressions().get(0));
     }
 
@@ -48,15 +49,17 @@ public class UrlDecodeFunction implements ValueParser {
         if (stringObj == null) {
             return null;
         }
+
         String string = OperatorTools.parseString(stringObj);
         if (string == null) {
             return null;
         }
 
         try {
-            return URLDecoder.decode(string, StandardCharsets.UTF_8.toString());
+            return URLEncoder.encode(string, StandardCharsets.UTF_8.toString());
         } catch (Exception e) {
             return null;
         }
     }
+
 }
